bug fix in mergesort, modified loader, created Timer
There was a small bug in mergesort that made it crash.
The loader is now more powerful, it will load an array of js files and
will call a callback (because of some issues of asynchronous loading).
Created a Timer "class", which can time any algorithm and compare two
algorithms.